Rename outer docgenerator to doctools
To make it easier for people to understand the relationship between the musicxmldoc.json/db.sqlite3 files and the (mnx)docgenerator project, the outer folder is renamed to doctools.

Doctools README.md is updated.

Main README.md file is updated to include requirement to sign license agreements before PRs and acknowledge Michael Good's role in MusicXML